Conflicts:
	src/server/scripts/Spells/spell_dk.cpp
	src/server/scripts/Spells/spell_paladin.cpp
	src/server/scripts/Spells/spell_priest.cpp
	src/server/scripts/Spells/spell_shaman.cpp
	src/server/scripts/Spells/spell_warrior.cpp

+++ src/server/game/Entities/Creature/Creature.h
@@ -247,7 +247,7 @@ struct CreatureData
     CreatureData() : dbData(true) {}
     uint32 id;                                              // entry in creature_template
     uint16 mapid;
-    uint16 phaseMask;
+    uint32 phaseMask;
     uint32 displayid;
     int8 equipmentId;
     float posX;
